First commit; working code, with bad profile
The diffraction ring currently does not correct for the fact
that area is not preserved upon stereographic projection. The
code also has the number of bins hardcoded.

To correct the diffraction ring, more bins mauy be necessary.
The diffraction ring can be thought of as two rings with different
centres and radii. By looking at the He-alpha width one can
define these two circles and then select all bins that lie between
them.