Fixed diffraction ring; made bin number a variable
The diffraction ring is now calculated between two values of
wavelength; these wavelengths have currently been selected based
on the width of a Zn spectrum from the literature. The number
of bins 'num_bins' used in the final image is now a variable; it
is no longer hard coded and can be changed as the user requires.

The calculation of phi is currently incorrect. The new diffraction
ring method has not yet been applied to the powder.